Rewrite 2023-17
Instead of storing the path history in each node, the nodes are instead the position and the direction of arrival. The neighbors of a node are, after turning, the positions after some steps depending on the part -- [1,3] for part 1, [4,10] for part 2.

Apparently, this makes the node graph shallower, thus the speedup.